implemented externally-facing capmix module

+++ src/c/lib/memory.c
@@ -17,7 +17,7 @@ void capmix_memory_init()
 			memory.section[s].buffer[o] = capmix_Unset;
 }
 
-int capmix_memory_erase(capmix_Addr addr, size_t len)
+int capmix_memory_erase(capmix_Addr addr, int len)
 {
 	capmix_Coord coord = capmix_addr_coord(addr);
 	if( coord.section > N_MEMSECTION || len + coord.offset > N_MEMBUF )
@@ -35,7 +35,7 @@ uint8_t *capmix_memory_get(capmix_Addr addr)
 	return mem;
 }
 
-int capmix_memory_set(capmix_Addr addr, uint8_t *data, size_t len)
+int capmix_memory_set(capmix_Addr addr, uint8_t *data, int len)
 {
 	capmix_Coord coord = capmix_addr_coord(addr);
 	if( coord.section > N_MEMSECTION || len + coord.offset > N_MEMBUF )
